+++ server/src/main/java/com/hobbyt/domain/privatehome/dto/response/PrivateHomePostLikeResponse.java
@@ -1,4 +1,4 @@
-package com.hobbyt.domain.privatehome.dto;
+package com.hobbyt.domain.privatehome.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -9,21 +9,23 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
-public class PrivateHomeSaleLikeResponse {
+public class PrivateHomePostLikeResponse {
 	private Boolean hasNext;
-	private List<SaleCard> cards;
+	private List<PostCard> cards;
 
 	@Getter
 	@NoArgsConstructor
-	public static class SaleCard {
-		private Long saleLikeId;
-		private Long saleId;
+	public static class PostCard {
+		private Long postLikeId;
+		private Long postId;
+		private Long postWriterId;
+		private String postWriterNickname;
+		private String profileImage;
 		private String title;
 		private String content;
 		private String thumbnailImage;
 		private long viewCount;
 		private long likeCount;
 		private LocalDateTime createdAt;
-		private String profileImage;
 	}
 }
