+++ server/src/main/java/com/hobbyt/domain/privatehome/dto/response/PrivateHomePostResponse.java
@@ -1,4 +1,4 @@
-package com.hobbyt.domain.privatehome.dto;
+package com.hobbyt.domain.privatehome.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -9,23 +9,21 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
-public class PrivateHomePostLikeResponse {
+public class PrivateHomePostResponse {
 	private Boolean hasNext;
-	private List<PostCard> cards;
+	private List<PostCard> posts;
 
 	@Getter
 	@NoArgsConstructor
 	public static class PostCard {
-		private Long postLikeId;
-		private Long postId;
-		private Long postWriterId;
-		private String postWriterNickname;
-		private String profileImage;
+		private Long writerId;
+		private Long id;
 		private String title;
 		private String content;
 		private String thumbnailImage;
 		private long viewCount;
 		private long likeCount;
+		private Boolean isPublic;
 		private LocalDateTime createdAt;
 	}
 }
