+++ server/src/main/java/com/hobbyt/domain/privatehome/dto/response/PrivateHomeSaleLikeResponse.java
@@ -1,4 +1,4 @@
-package com.hobbyt.domain.privatehome.dto;
+package com.hobbyt.domain.privatehome.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -9,21 +9,21 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
-public class PrivateHomePostResponse {
+public class PrivateHomeSaleLikeResponse {
 	private Boolean hasNext;
-	private List<PostCard> posts;
+	private List<SaleCard> cards;
 
 	@Getter
 	@NoArgsConstructor
-	public static class PostCard {
-		private Long writerId;
-		private Long id;
+	public static class SaleCard {
+		private Long saleLikeId;
+		private Long saleId;
 		private String title;
 		private String content;
 		private String thumbnailImage;
 		private long viewCount;
 		private long likeCount;
-		private Boolean isPublic;
 		private LocalDateTime createdAt;
+		private String profileImage;
 	}
 }
