+++ server/src/test/java/com/hobbyt/util/TestUtil.java
@@ -6,15 +6,15 @@ import java.time.LocalDateTime;
 
 import org.springframework.mock.web.MockMultipartFile;
 
-import com.hobbyt.domain.member.dto.request.ProfileRequest;
 import com.hobbyt.domain.member.dto.request.UpdateMyInfoRequest;
 import com.hobbyt.domain.member.dto.request.UpdatePassword;
 import com.hobbyt.domain.member.dto.response.MyInfoResponse;
-import com.hobbyt.domain.member.dto.response.ProfileResponse;
 import com.hobbyt.domain.member.entity.Authority;
 import com.hobbyt.domain.member.entity.Member;
 import com.hobbyt.domain.member.entity.Recipient;
 import com.hobbyt.domain.member.entity.Views;
+import com.hobbyt.domain.privatehome.dto.request.ProfileRequest;
+import com.hobbyt.domain.privatehome.dto.response.ProfileResponse;
 import com.hobbyt.global.entity.Account;
 import com.hobbyt.global.entity.Address;
 import com.hobbyt.global.security.member.MemberDetails;
